requirements.txt
streamlit
numpy
scikit-learn
sentence-transformers
openai
langchain-text-splitters
PyPDF2
python-docx
tavily-python
requests
